Feature/cron job feature

+++ social-pro-be/src/main/java/com/spring/social_pro/backend/dto/request/payment/PaymentGenerateRequest.java
@@ -3,12 +3,14 @@ package com.spring.social_pro.backend.dto.request.payment;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class PaymentGenerateRequest {
-    Long amount;
+    BigDecimal amount;
     String gate;
 }
